requirements.txt
Django==1.11.2
django-extra-views==0.9.0
django-test-without-migrations==0.6
